Add limit pushdown rule (#55)
This PR adds a custom catalyst rule for doing limit pushdown for Delta Sharing queries. When the rule is applied, the limit in SQL queries will be pushed down and sent to the server as `limitHints`. e.g. `SELECT * FROM table LIMIT 100`

+++ server/src/main/scala/io/delta/standalone/internal/DeltaSharedTableLoader.scala
@@ -118,7 +118,7 @@ class DeltaSharedTable(
   def query(
       includeFiles: Boolean,
       predicateHits: Seq[String],
-      limitHint: Option[Int]): (Long, Seq[model.SingleAction]) = withClassLoader {
+      limitHint: Option[Long]): (Long, Seq[model.SingleAction]) = withClassLoader {
     // TODO Support `limitHint`
     val snapshot = deltaLog.snapshot
     validateDeltaTable(snapshot)
